Reduce test duplication
